+++ apps/api-gateway/src/models/marketing_campaign.py
@@ -1,7 +1,7 @@
 """
 营销活动模型
 """
-from sqlalchemy import Column, String, Integer, Float, DateTime, JSON, Date, Text
+from sqlalchemy import Column, String, Integer, Float, DateTime, JSON, Date, Text, Numeric
 from sqlalchemy.sql import func
 import uuid
 
@@ -21,13 +21,13 @@ class MarketingCampaign(Base):
     start_date = Column(Date)
     end_date = Column(Date)
 
-    budget = Column(Float, default=0.0, comment="预算（元）")
-    actual_cost = Column(Float, default=0.0, comment="实际花费（元）")
+    budget = Column(Numeric(12, 2), default=0.0, comment="预算（元）")
+    actual_cost = Column(Numeric(12, 2), default=0.0, comment="实际花费（元）")
 
     # 效果指标
     reach_count = Column(Integer, default=0, comment="触达人数")
     conversion_count = Column(Integer, default=0, comment="转化人数")
-    revenue_generated = Column(Float, default=0.0, comment="带来营收（元）")
+    revenue_generated = Column(Numeric(12, 2), default=0.0, comment="带来营收（元）")
 
     target_audience = Column(JSON, comment="目标受众配置")
     description = Column(Text)
